Allow hbbft to proceed to the next round, with protocol extensions
The HoneyBadgerBFT paper is a bit light on the details of how it moves
to the next round. Indeed, the paper assumes transactions are filtered,
reordered or retained in the buffer via some undefined mechanism.

This commit implements a mechanism to do that, as well as a mechanism
for proving f+1 nodes all agreed on this block.

When a set of transactions has been agreed on, the hbbft protocol
returns `{result, {transactions, [binary()]}}`. The application then has
an opportunity to filter/reorder/retain any transactions it chooses. It
then calls finalize_round/3 with the set of transactions it wants the
protocol to *delete* from its buffer as well as a binary representing
the block over which it wants a threshold signature.

The protocol removes the listed transactions from its buffer and then
goes about trying to obtain a signature over the block. Assuming f+1
other nodes also performed the same block construction, a threshold
signature can be obtained and the block can be shown to have majority
agreement.