Fix a bug in RBC and a bug in ACS
RBC would ignore any VAL messages it saw once it saw an ECHO. This was
incorrect and it prevented this instance of RBC ever sending an ECHO on
its own behalf, which had the chance of preventing consensus.

ACS had an incorrect assumption about the termination condition. It
assumed that BBAs would only converge after the corresponding RBC. This
is incorrect and ACS must wait for all the RBCs that have a
corresponding BBA that returned 1 to themselves return. ACS has been
refactored to account for this.