Add a event stamp function for attaching causality information to RBC
To provide some semblence of causality we allow the user of hbbft to
provide a 'stamp fun' that allows them to attach some causality
information to the Reliable Broadcast payload. This can be a timestamp,
a monotonic counter, a vector clock, whatever. The resulting 'stamp' is
encrypted along with the other transactions and is thus subject to the
consensus rules around agreeing on transactions.