Allow ACS to return a result and messages to send simultaneously
After observing cases where all the required BBA and ACS instances
completed, but ACS itself did not return. I believe this is caused by
the final BBA completing, returning the final 1 that puts us over the
N-F threshold. However, the current code path assumes that the final BBA
won't be the final one to return 1, and also assumes that ACS can send a
message, or return a result, but not both.

This change allows ACS to return a result AND send messages, and updates
HBBFT itself to handle that return value. This along with the prior
change to ACS to check for a return in the RBC handling code, should
catch the missing termination conditions.