Validate ciphertexts only once and store them in the state
Instead of reconstructing and revalidating TPKE ciphertexts on-demand, we
construct and validate them once and store them in the state. If the key
is not in the state we know it was invalid and we have to disregard that
ACS result.

Additionally implement and use helper functions for
serializing/deserializing TPKE ciphertexts.

Finally, remove some legacy compatability code for untagged shares.