Update to new TPKE API for verified ciphertexts
TPKE has been updated to track the validity of the ciphertext inside the
ciphertext value itself and not check it for each invocation of
verify_share/combine_share/decrypt. This means that, to avoid
reverifying the ciphertext every time we need it, instead we cache the
verified ciphertexts in the state.

This all leads to a significant performance improvement as we can avoid
a lot of redundant ciphertext validations.

Additionally, clean up some specs and configure xref to not warn about
unused functions.