Retain our proposal and stamp so we don't have to decrypt them later
A share of RBC (and thus of ACS) can only complete if we've seen 2f+1
READY messages (and then f+1 ECHO messages). Thus we should, assuming
less than f+1 byzantine nodes, be able to have confidence we will get
f+1 decryption shares if our proposal gets included in the ACS result.

Doing this allows us to avoid verifying at least f+1 shares and the
lagrange interpolation over the valid shares to decrypt the proposal we
already know the contents of.