Improve status output and track tpke failures
The current HBBFT status output omits potentially important information
about *which* ACS results completed, if we got any invalid (or
unvalidated) decryption shares and which proposals we've been able to
decrypt.

In addition, if somehow share recombination fails, or the result of
combining the decryption shares does not allow us to decrypt the
proposal, that is very important to be able to see, so those are now
tracked as well.